Unary and binary expressions. Heavy code reformatting
Support for unary expression like prefix/postfix increments.
Suports for basic math operators.
Reformatting the BisonGrammar file indentation.
IMPORTANT: Now expressions should end with a SEMICOLON, the optional
semicolon was introducing a lot of conflicts. I rather enforce the
semicolon while finishing the rest of the statements now, and deal with
the optional semicolon in the future.